Change comments, make readability changes, update SQL variables, and add SignupServlet
